Create Assignment 5
2. INTRODUCTION:
Weather balloons are used to gather temperature and pressure data at various altitudes in the atmosphere. These
balloons are usually filled with helium, and this allows the balloon to rise (helium is less dense than air).
However, as the balloon rises, the surrounding air becomes less dense. As the densities of the helium and air
approach one another, the balloon’s rise slows and it reaches a point of equilibrium. Since the heat of the day
causes the helium in the balloon to expand and become less dense, the balloon is at a higher altitude during the
day than it is at night. For a particular weather balloon, this process of rising and falling altitudes can be
approximated with the following polynomial equation:
𝑎𝑙𝑡𝑖𝑡𝑢𝑑𝑒 = −0.15𝑡
4 + 12.5𝑡
3 − 280𝑡
2 + 4100𝑡 + 230
where altitude is in meters and t (time) is in hours. The corresponding polynomial model for the velocity (m/h) of
the weather balloon is
𝑣𝑒𝑙𝑜𝑐𝑖𝑡𝑦 = −0.60𝑡
3 + 37.5𝑡
2 − 560𝑡 + 4100
ASSIGNMENT:
Write a C program that will print a table of the altitude and the velocity for the weather balloon in meters and
meters per second, respectively. Prompt the user for the start time (in hours after launch, where launch = 0 hours),
the end time, and the increment in time between lines of the table. The equations are accurate only for the first 42
hours after launch, so check the user input to make sure that it stays within the proper bounds. If the user input does
not stay within the proper bounds, ask the user to re-enter any information that is not acceptable until acceptable
inputs are provided (hint: use while). To calculate the altitude and velocity, use a for loop. Print the peak altitude
and its corresponding time after the table.
Use the “Code Sections.c” template on Blackboard to write your code.
Your program output will look like the illustration shown below. Use your PC’s cursor to determine the
horizontal and vertical spacing for the output format. 30 points
2
SAMPLE OUTPUT FORMAT:
********************************************
 ALTITUDE AND VELOCITY OF WEATHER BALLOON
Enter start time in hours: x
Enter final time in hours: x
Enter increment between lines in hours: x
RESULTS
TIME (h) ALTITUDE (m) VELOCITY (m/s)
xx.xx xxxxx.x x.xx
xx.xx xxxxx.x x.xx
 . . .
 . . .
 . . .
PEAK ALTITUDE: xxxxx.x m
TIME: xx.xx h