Create Assignment 6
INTRODUCTION:
Oil exploration is being conducted by drilling holes and blasting the ground with specific amounts of dynamite to
generate seismic readings that can be interpreted. To obtain useful seismic readings, a specific ratio between the
amount of dynamite and the depth of the hole is needed. The ideal powder charge to depth-of-hole ratio is 1:3.
For blasting operations, each stick of dynamite will be 2 feet long and will weigh 6 pounds. Ideally, we would
use the exact amount of powder specified; however, in reality, useful readings aren’t generated when part of a
stick of dynamite protrudes above the hole. Therefore, we will compute the actual powder charge in 6-pound
increments and use fewer sticks than necessary, if needed.
A sample hand calculation follows below:
Hole depth = 85 feet
Ideal powder-to-hole depth ratio = 85/3 = 28.333333 feet
Ideal number of sticks = 28.333333/2 = 14.166667
Ideal powder charge = 14.166667*6 = 85 pounds
Actual number of sticks = floor(14.166667) = 14
Actual charge = 14*6 = 84 pounds
An input data file called explore.txt contains exploration information for each well site. The first record line of
the input file is the control number, which defines the number of sites to be read into the program. Each
succeeding record line has two columns, where the first column is the well site number and the second column is
the hole depth of the well in feet. The input file is comma delimited.
ASSIGNMENT:
Write a C program to read the information for each hole from the explore.txt data file (one point at a time). Using
the hole depth, compute the ideal powder-to-hole depth ratio, the ideal number of sticks, the ideal powder charge,
the actual powder charge, and the actual number of sticks. Your program will produce a table as the report. Print
the table to the computer screen and to an output file called blasting.txt.
Once the program is working, modify the program to account for any hole depth that is less than 30 feet. Instead
of printing the normal column items, for any hole that is less than 30 feet in depth, print the well site number, the
well depth, and then print a message that the hole is too shallow.