Create Assignment 7
An example of a simple resistor circuit is displayed below.
In such circuits, V1 is a voltage source (V) and R1 is a resistor (Ω). The current in these circuits may be
calculated using Ohm’s law:
𝐼 =𝑉/𝑅
The power dissipated across the resistor is
𝑃 = (𝐼^2)*R
In the above equations, I is the current in amps (A), V is the voltage in volts (V), R is the resistance in ohms (Ω),
and P is the power in Watts (W).
An input data file called circuits contains voltage and resistance information for a set of simple resistor circuits.
The first record line of the input file is the control number, which defines the number of circuits to be read into
the program. Each succeeding record line has two columns, where the first column is the voltage and the second
column is the resistance. The input file is comma delimited.
ASSIGNMENT:
Write a C program that will read the required values from the input file. Using the voltage and resistance,
compute the current in each circuit and the power dissipated by the resistor in each circuit. Write a user-defined
function to compute the current and a user-defined function to compute the power. Your program will produce a
report in tabular format. Print the table to the computer screen and to an output file called levels.
Once the program is working, modify the program to account for any current that is greater than 0.1 amp. Instead
of printing the normal column items, for any circuit with current greater than 0.1 amp, print the voltage, the
resistance, and then print a message that the current may be fatal. The output format is shown on page 2.