Create Assignment 8
Write a C program that will read the required values from the input file into two one-dimensional arrays: one array
for the initial length and one array for the final length. Calculate the strain based on the length values, and use this
strain to calculate the stress. Print the strain, the stress in ksi, and which regime the strain corresponds to, both to
the screen and to an output data file called stress_strain